+++ src/domain/collaboration/aspect/views/AspectDialogSection.ts
@@ -1,4 +1,5 @@
 export enum AspectDialogSection {
   Dashboard = 'dashboard',
+  Share = 'share',
   Settings = 'settings',
 }
