+++ src/domain/collaboration/aspect/views/AspectRoute.tsx
@@ -2,6 +2,7 @@ import React, { FC } from 'react';
 import { Navigate, Route, Routes, useNavigate } from 'react-router-dom';
 import AspectDashboardPage from '../pages/AspectDashboardPage';
 import AspectSettingsPage from '../pages/AspectSettingsPage';
+import AspectSharePage from '../pages/AspectSharePage';
 import { AspectDialogSection } from './AspectDialogSection';
 import { AspectLayoutHolder } from './AspectLayoutWithOutlet';
 
@@ -19,6 +20,7 @@ const AspectRoute: FC<AspectRouteProps> = () => {
       <Route path="/" element={<AspectLayoutHolder />}>
         <Route index element={<Navigate replace to={AspectDialogSection.Dashboard} state={{ keepScroll: true }} />} />
         <Route path={AspectDialogSection.Dashboard} element={<AspectDashboardPage onClose={onClose} />} />
+        <Route path={AspectDialogSection.Share} element={<AspectSharePage onClose={onClose} />} />
         <Route path={AspectDialogSection.Settings} element={<AspectSettingsPage onClose={onClose} />} />
       </Route>
     </Routes>
